huangshiwei4@huawei.com

+++ services/uripermmgr/src/uri_permission_manager_stub_impl.cpp
@@ -183,7 +183,7 @@ void UriPermissionManagerStubImpl::RevokeAllUriPermissions(int tokenId)
         for (auto iter = uriMap_.begin(); iter != uriMap_.end();) {
             auto& list = iter->second;
             for (auto it = list.begin(); it != list.end();) {
-                if (it->targetTokenId == tokenId || it->fromTokenId == tokenId) {
+                if (it->targetTokenId == static_cast<int>(tokenId) || it->fromTokenId == static_cast<int>(tokenId)) {
                     HILOG_INFO("Erase an info form list.");
                     uriLists[it->targetTokenId].emplace_back(iter->first);
                     list.erase(it++);
